bugfix
Change thresholds rep from n.parameters to max.length. This is what it
used to be and I have no memory of why I changed it. If it is
n.parameters it doesn’t fill the list with NAs and the as.matrix fails
and the whole thing collapses. Use Jinho’s multidim as test for this if
considering changing back.